Rebuild the test/test_files/airtable.browser.js
- Tested that `node-fetch` replacement actually works by running the
`test/test_files/index.html`
- In order to test againast the latest changes I regenerated the
browserifyed js with `node_modules/grunt/bin/grunt browserify`.
- I copied the built file to the test directly.
- I'm unclear what the process is for when `airtable.browser.js` should
be updated. I didn't check in the built version, because I assume that
is only updated when I release is done, but I'm not clear what the
process is. That would be something that would be great to clarify in
the README.