Add polyfills for older browser versions
- Add AbortController polyfill for Safari 10 - 11 and Firefox 52 - 56
- Was going to add fetch polyfill but it is only missing from Safari 10.0. It is in Safari 10.1 and all
other browsers that need to be supported. All Safari before 10.1 is only
0.26% of all browsers usage. It is not possible to test Safari 10.0 on
Browserstack and the complexity and size cost of adding the polyfill
for this one minor version is large, so it was skipped it.

Including the one polyfill brings the size of the built package to 232K up
from 220K.